requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.3.0
gcloud==0.18.3
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
jws==0.1.3
oauth2client==4.1.3
passlib==1.7.4
phonenumbers==8.13.4
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
Pyrebase4==4.6.0
python-jwt==2.0.1
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.14
